wip/removed dataclasses + added db models + added db manager
