+++ curueno_porma_warbot/warbot.py
@@ -1,0 +1,1 @@
+from curueno_porma_warbot.manager import DBManagerfrom curueno_porma_warbot.twitter_handler import TwitterAPIClientfrom curueno_porma_warbot.utils.log import ClassWithLoggerclass CuruenoPormaWarbot(ClassWithLogger):    def __init__(self, *args, **kwargs):        super().__init__(*args, **kwargs)        self.tw_client = TwitterAPIClient()        self.db_manager = DBManager()        self._end = False    def __enter__(self):        self.logger.info("Start CuruenoPormaWarbot")        while self._end is not True:            pass    def __exit__(self, exc_type, exc_val, exc_tb):        self.logger.info("Exit CuruenoPormaWarbot")
