--- src/warbot.py
@@ -1,1 +1,0 @@
-from src.managers import TownManager, EventManagerfrom src.twitter_handler import TwitterAPIClientfrom src.utils.log import ClassWithLoggerclass CuruenoPormaWarbot(ClassWithLogger):    def __init__(self, *args, **kwargs):        super().__init__(*args, **kwargs)        self.tw_client = TwitterAPIClient()        self.town_manager = TownManager()        self.event_manager = EventManager()        self.town_manager.init_town_list()        self.event_manager.init_event_list()        self._end = False    def __enter__(self):        self.logger.info("Starting Warbot!")        while self._end is not True:            pass    def __exit__(self, exc_type, exc_val, exc_tb):        self.logger.info("Ending Warbot!")
